C functions

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,14 +9,14 @@
 int main(void)
 {
 int count = 0;
-double a = 0, b = 1, c = 0;
+unsigned long long int a = 0, b = 1, c = 0;
 
 for (count = 0; count <= 98; count++)
 {
 	c = a + b;
 	a = b;
 	b = c;
-	printf("%.0lf", c);
+	printf("%llu", c);
 	if (count != 98)
 		printf(", ");
 }
